# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Cadrly API
  version: 0.1.0
  description: Приложение для обмена коллекционными карточками
servers:
  - description: Test server
    url: http://localhost:8080/api
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Регистрация нового пользователя
      description: |
        Принимает email, имя пользователя и пароль.
        Если email и пароль валидны, отправляет код подтверждения.
        Возвращает временный токен для верификации.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: 'user@example.com'
                  description: Email пользователя
                username:
                  type: string
                  minLength: 3
                  maxLength: 25
                  example: 'Коллекционер_123'
                  description: 'Имя пользователя (3-25 символов)'
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: 'Password123!'
                  description: 'Пароль (минимум 8 символов)'
              required: [email, username, password]
      responses:
        200:
          description: Код подтверждения отправлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  tempToken:
                    type: string
                    example: 'temp_abc123'
                  message:
                    type: string
                    example: 'Код подтверждения отправлен на email'
        400:
          description: Неверный формат данных
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Некорректный формат введенных данных'
        409:
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Пользователь с таким email уже существует'
        500:
          $ref: '#/components/responses/ServerError'
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Вход в систему
      description: |
        Аутентификация пользователя по email и паролю.
        Если данные введены верно - устанавливает сессионную cookie.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: 'user@example.com'
                password:
                  type: string
                  format: password
                  example: 'SecurePass123!'
              required: [email, password]
      responses:
        200:
          description: Успешный вход
          headers:
            Set-Cookie:
              schema:
                type: string
                example: 'session_id=abcde12345; HttpOnly; Path=/; Secure'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Неверный формат email
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Неверный фортмат email'

        401:
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Неверный email или пароль'
        500:
          $ref: '#/components/responses/ServerError'
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Запрос кода для сброса пароля
      description: Отправляет код подтверждения для сброса пароля, если email существует в системе
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: 'user@example.com'
              required: [email]
      responses:
        200:
          description: Код подтверждения отправлен на email
          content:
            application/json:
              schema:
                type: object
                properties:
                  resetToken:
                    type: string
                    example: 'reset_xyz789'
                  message:
                    type: string
                    example: 'Код подтверждения отправлен на email'
        400:
          description: Неверный формат данных
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Некорректный формат emaila'
        404:
          description: Пользователь с указанным email не найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Пользователь с таким email не зарегистрирован'
        500:
          $ref: '#/components/responses/ServerError'
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Сброс пароля
      description: Устанавливает новый пароль после подтверждения кода
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resetToken:
                  type: string
                  example: 'reset_xyz789'
                code:
                  type: string
                  pattern: "^\\d{6}$"
                  example: '654321'
                newPassword:
                  type: string
                  format: password
                  minLength: 8
                  example: 'NewSecurePass123!'
              required: [resetToken, code, newPassword]
      responses:
        200:
          description: Пароль изменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Пароль успешно изменен'
        400:
          description: Неверный код или токен
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Неверный код подтверждения'
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    Achievement:
      type: object
      properties:
        achievement_ID:
          type: integer
          example: 101
          description: 'Уникальный идентификатор достижения'
        name:
          type: string
          example: 'Мастер коллекционирования'
          description: 'Название достижения'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/achievements/master.png'
          description: 'Ссылка на изображение'
        description:
          type: string
          example: 'Соберите 10 полных коллекций'
          description: 'Описание достижения'
        isUnlocked:
          type: boolean
          example: false
          description: 'Статус получения'
      required:
        - achievement_ID
        - name
        - imageURL
        - description
        - isUnlocked
    Notification:
      type: object
      properties:
        notification_ID:
          type: integer
          example: 201
          description: 'Уникальный идентификатор уведомления'
        user:
          $ref: '#/components/schemas/User'
          description: 'Пользователь, которому отправлено уведомление'
        message:
          type: string
          example: 'Ваша карта была успешно обменяна'
          description: 'Текст уведомления'
        links:
          type: array
          items:
            type: string
            format: url
          example: ['https://example.com/exchange-details']
          description: 'Ссылка на подробности обмена'
        notificationDateTime:
          type: string
          format: date-time
          example: '2024-05-20T15:30:00Z'
          description: 'Дата и время уведомления'
      required:
        - notification_ID
        - user
        - message
        - notificationDateTime
...
    Card:
      type: object
      properties:
        card_ID:
          type: integer
          example: 501
          description: 'Уникальный идентификатор карты'
        name:
          type: string
          example: 'Ледяной феникс'
          description: 'Название карты'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/cards/pheonix.png'
          description: 'Ссылка на изображение'
        description:
          type: string
          example: 'Ледяной Феникс — это мифическое существо, воплощающее силу зимы и вечного обновления. Его перья сверкают как морозный утренний иней, а глаза светятся холодным синим светом.'
          description: 'Описание карты'
        rarity:
          type: string
          enum: ['Обычная', 'Редкая', 'Эпическая', 'Легендарная', 'Уникальная']
          example: 'Редкая'
          description: 'Редкость карты'
        min_price:
          type: integer
          example: 50
          description: 'Минимальная цена - используется для разбора карточки'
        theme:
          type: string
          nullable: true
          example: 'Мифическое существо'
          description: 'Тематика, на которую была сгенерирвана карта (только для уникальных карт)'
        isGenerated:
          type: boolean
          example: false
          description: 'Флаг показывающий сгенерированная ли карта (true для уникальных, false для остальных)'
      required:
        - card_ID
        - name
        - imageURL
        - rarity
        - min_price
        - isGenerated
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
          description: 'Уникальный идентификатор'
        username:
          type: string
          example: 'Коллекционер_123'
          description: 'Имя пользователя'
        email:
          type: string
          format: email
          example: 'user@example.com'
          description: 'Электронная почта'
        password:
          type: string
          format: password
          writeOnly: true
          description: 'Пароль'
        favoriteCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Избранные карты'
        onChange:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Карты на обмен'
        inventoryCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Карты в инвентаре'
        balance:
          type: integer
          example: 2500
          description: 'Баланс'
        avatar_url:
          type: string
          format: url
          example: 'https://example.com/avatars/user1.png'
          description: 'Аватар пользователя'
        achievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
          description: 'Достижения'
        favoriteAchievements:
          type: array
          items:
            $ref: '#/components/schemas/Achievement'
          description: 'Избранные достижения'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
          description: 'Уведомления'
      required:
        - user_id
        - username
        - email
        - password
        - balance
        - inventoryCards
        - achievements
...
  responses:
...
    ServerError:
      description: Что-то пошло не так
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Что-то пошло не так, попробуйте позже'
              code:
                type: string
                example: 'INTERNAL_SERVER_ERROR'
              timestamp:
                type: string
                format: date-time
                example: '2024-05-20T12:00:00Z'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT